@@ -1,16 +1,19 @@
 package com.hencoder.hencoderpracticedraw4;
 
+
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewStub;
 
+import androidx.annotation.LayoutRes;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
 public class PageFragment extends Fragment {
-    @LayoutRes int sampleLayoutRes;
+    @LayoutRes
+    int sampleLayoutRes;
     @LayoutRes int practiceLayoutRes;
 
     public static PageFragment newInstance(@LayoutRes int sampleLayoutRes, @LayoutRes int practiceLayoutRes) {
